Ship speed changed to suite N950
The divider that limits the ship speed is 5 for Harmattan, 15 for Maemo.
The way the change was done it also means that with Harmattan 5 degree
tilt is enough ti trigger move, whereas Maemo needs 15 degrees. Need to
test if this feels good.
Again, no proper ifdefs, just commenting Maemo stuff out and writing
Harmattan stuff in. *Must* ifdef properly before merging to master.