Bit faster
Speed is better now, still slower than on Maemo, but roughly what
it was before on Harmattan.
Bug: speed sopetimes drops for a brief moment. This seems to be
caused by system slowing because of lots of apps running. (Appeared
with 7 browser windows and three other apps on the background, but
not with no apps on the background.) The bug is not caused by the
change to timer-triggered sensor reading, and seemed even worse
without it.