Runs on sailfish emulator
Outcommented installing .desktop, launch icon and resource policy.

Even the cover seems to work.

Issues:

Game is in portrait mode - not all of the menu is visible.
Scene size is hardcoded to that of the N9/N950 screen.
Emulator has no vibra device ;-)
Rotationsensor interface not granted. Probaply because there is no
accelometer in the emulator ;-)

As obvious, the game cannot be played in the emulator as it is
accelometer controlled...

No launcher icon, as the icon and .desktop file are not installed.